資源の保留解除、譲渡を資源量単位で行えるよう修正

+++ SangokuKmy/Models/Data/Entities/CharacterItem.cs
@@ -78,7 +78,7 @@ namespace SangokuKmy.Models.Data.Entities
 
     [Column("resource")]
     [JsonProperty("resource")]
-    public ushort Resource { get; set; }
+    public int Resource { get; set; }
   }
 
   public enum CharacterItemStatus : short
